Restructured GMSH IO as separate sub module. Expose GhostMode and CellType from C++ + moving mesh to separate module
