+++ python/demo/poisson/demo_poisson.py
@@ -74,19 +74,17 @@
 #
 # First, the :py:mod:`dolfinx` module is imported: ::
 
-import matplotlib.pyplot as plt
-import numpy as np
-from mpi4py import MPI
-from petsc4py import PETSc
-
 import dolfinx
 import dolfinx.plotting
+import matplotlib.pyplot as plt
+import numpy as np
 import ufl
 from dolfinx import DirichletBC, Function, FunctionSpace, RectangleMesh, solve
-from dolfinx.cpp.mesh import CellType
 from dolfinx.fem import locate_dofs_topological
 from dolfinx.io import XDMFFile
-from dolfinx.mesh import locate_entities_boundary
+from dolfinx.mesh import CellType, GhostMode, locate_entities_boundary
+from mpi4py import MPI
+from petsc4py import PETSc
 from ufl import ds, dx, grad, inner
 
 # We begin by defining a mesh of the domain and a finite element
@@ -100,7 +98,7 @@ from ufl import ds, dx, grad, inner
 mesh = RectangleMesh(
     MPI.COMM_WORLD,
     [np.array([0, 0, 0]), np.array([1, 1, 0])], [32, 32],
-    CellType.triangle, dolfinx.cpp.mesh.GhostMode.none)
+    CellType.triangle, GhostMode.none)
 
 V = FunctionSpace(mesh, ("Lagrange", 1))
