+++ python/dolfinx/mesh/__init__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2020 Jørgen S. Dokken
+#
+# This file is part of DOLFINX (https://www.fenicsproject.org)
+#
+# SPDX-License-Identifier:    LGPL-3.0-or-later
+"""Tools for creating and manipulating meshes for the finite element method"""
+
+from dolfinx.cpp.mesh import CellType, Geometry, Topology, GhostMode
+
+from .mesh import MeshTags, create_mesh, create_meshtags, locate_entities, locate_entities_boundary, refine
+
+__all__ = [
+    "locate_entities", "locate_entities_boundary", "refine", "create_mesh", "create_meshtags", "MeshTags",
+    "Topology", "Geometry", "GhostMode", "CellType"]
