+++ python/setup.py
@@ -80,8 +80,10 @@ setup(name='fenics-dolfinx',
       packages=["dolfinx",
                 "dolfinx.fem",
                 "dolfinx.la",
+                "dolfinx.mesh",
                 "dolfinx.wrappers",
-                "dolfinx_utils.test"],
+                "dolfinx_utils.test",
+                "dolfinx.mesh.gmsh"],
       package_data={'dolfinx.wrappers': ['*.h']},
       ext_modules=[CMakeExtension('dolfinx.cpp')],
       cmdclass=dict(build_ext=CMakeBuild),
