+++ python/test/unit/io/test_xdmf_mesh.py
@@ -9,9 +9,9 @@ import os
 import numpy as np
 import pytest
 from dolfinx import UnitCubeMesh, UnitIntervalMesh, UnitSquareMesh, cpp
-from dolfinx.cpp.io import perm_gmsh
+from dolfinx.mesh.gmsh import perm_gmsh, ufl_mesh_from_gmsh
 from dolfinx.cpp.mesh import CellType
-from dolfinx.io import XDMFFile, ufl_mesh_from_gmsh
+from dolfinx.io import XDMFFile
 from dolfinx.mesh import create_mesh
 from dolfinx_utils.test.fixtures import tempdir
 from mpi4py import MPI
